feat(character): update request with new field
