+++ server.js
@@ -1,6 +1,5 @@
 import express from 'express';
 import mongoose from 'mongoose';
-
 import charactersRouter from './routes/characters.js';
